server/pyproject.toml
[tool.poetry]
name = "web-fall-2022"
version = "0.0.1"
description = "gRPC server"
authors = ["Crewhee <tony.sudov@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "web_fall_2022"}]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.49.1"
grpcio-tools = "^1.49.1"
grpcio-testing = "^1.49.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
flake8-bugbear = "^22.9.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"